+++ DeterminantApp/src/main/java/com/app/matrix/Matrix.java
@@ -6,13 +6,13 @@ import org.springframework.stereotype.Component;
 import org.springframework.beans.factory.annotation.Autowired;
 
 public class Matrix {
-	int[][] matrix;
+	double[][] matrix;
 	int rows;
 	int columns;
 	int rowCounter;
 
 	public Matrix(int d) {
-		this.matrix = new int[d][d];
+		this.matrix = new double[d][d];
 		this.rows = d;
 		this.columns = d;
 		this.rowCounter = 0;
@@ -26,7 +26,7 @@ public class Matrix {
 		return columns;
 	}
 	
-	public int getElement(int row, int col) {
+	public double getElement(int row, int col) {
 		return matrix[row][col];
 	}
 	
@@ -36,7 +36,7 @@ public class Matrix {
 		return false;
 	}
 	
-	public void addRow(int[] newRow) throws Exception {
+	public void addRow(double[] newRow) throws Exception {
 		if(newRow.length != columns)
 			throw new Exception("Error: Row must be of length " + columns);
 		if(rowCounter == 0) { 
@@ -50,7 +50,7 @@ public class Matrix {
 		}
 	}
 	
-	public void addColumn(int[] newCol, int colNum) throws Exception {
+	public void addColumn(double[] newCol, int colNum) throws Exception {
 		if(newCol.length != rows)
 			throw new Exception("Error: Column must be of length " + rows);
 		if(colNum > (columns - 1)) {
@@ -62,7 +62,7 @@ public class Matrix {
 	}
 	
 	public void printMatrix() {
-		for(int[] row: matrix) {
+		for(double[] row: matrix) {
 			System.out.print(Arrays.toString(row) + "\n");
 		}
 	}
